src/utils/process_files.py
from typing import List, Dict, Any
import os

# import pymupdf4llm
import json
import fitz

from pydantic import BaseModel, Field
from helper import logger
from utils.file_ops import save_file

from settings import Settings

from pdf_text_extractor.src import PDFTextExtractor, ExtractionMode


class Metadata(BaseModel):
    title: str = Field(default="")
    author: list[str] = Field(default_factory=list)
    affiliations: list[str] = Field(default_factory=list)
    subject: str = Field(default="")
    doi: str = Field(default="")
    year: str = Field(default="")
    month: str = Field(default="")
    keywords: list[str] = Field(default_factory=list)
    summary: str = Field(default="")


extractor = PDFTextExtractor(
    ExtractionMode.TABLE_IMAGE_LINKS, print_page_number=True, remove_header_footer=False
)


def process_uploaded_pdfs(
    file_paths: List[str], extract_metadata: bool = False
):  # -> List[Document]:
    """
    Process the uploaded PDFs. Extracts Markdown text from PDFs and saves it along with metadata
    """
    all_docs = []
    for file_path in file_paths:
        filename = os.path.basename(file_path).split(".")[0]
        folder = os.sep.join(file_path.split(os.sep)[:-1])
        text_folder = os.path.join(folder, filename)
        image_folder = os.path.join(text_folder, "images")
        os.makedirs(text_folder, exist_ok=True)
        # os.makedirs(image_folder, exist_ok=True)

        # markdown = pymupdf4llm.to_markdown(
        #     file_path,
        #     write_images=True,
        #     table_strategy="lines_strict",
        #     image_path=image_folder,  # Folder where images will be saved
        #     image_format="png",
        # )

        markdown = extractor.extract(file_path, image_folder=image_folder)

        # all_docs.extend(mark)
        save_file(os.path.join(text_folder, filename + ".md"), markdown)
        if extract_metadata:
            get_metadata(file_path, markdown)
        os.remove(file_path)
    logger.info("Successfully processed input PDFs")
    return all_docs


def get_metadata(file_path: str, markdown: str):
    filename = os.path.basename(file_path).split(".")[0]
    folder = os.sep.join(file_path.split(os.sep)[:-1])
    text_folder = os.path.join(folder, filename)
    with fitz.open(file_path) as doc:
        metadata = doc.metadata
        metadata["pages"] = doc.page_count
        metadata["toc"] = doc.get_toc()
    metadata["filename"] = filename
    with open(
        os.path.join("src", "utils", "metadata_prompt.txt"), "r", encoding="utf-8"
    ) as f:
        prompt_template = f.read().strip()

    prompt = prompt_template.format(text=markdown[:9000])
    response = run_llm(prompt)
    response = Metadata.model_validate_json(response)
    llm_metadata = json.loads(response.model_dump_json(indent=4, exclude_none=True))

    doc_metadata = merge_metadata(llm_metadata, metadata)
    doc_metadata = json.dumps(doc_metadata, indent=4)
    save_file(os.path.join(text_folder, filename + "_metadata.json"), doc_metadata)


def run_llm(prompt):
    messages = [("system", prompt)]
    response = Settings.llm.invoke(messages)

    return response.content


def merge_metadata(
    existing_metadata: Dict[str, Any], new_metadata: Dict[str, Any]
) -> Dict[str, Any]:
    """
    Merges new metadata into existing metadata.
    """

    def _merge_list_and_str(existing, new):
        if isinstance(existing, list) and isinstance(new, str):
            existing.append(new)
            return existing
        elif isinstance(existing, str) and isinstance(new, list):
            return list(set(existing.split(",")).union(set(new)))
        return existing

    for key, new_value in new_metadata.items():
        if key not in existing_metadata:
            existing_metadata[key] = new_value
            continue
        existing_value = existing_metadata[key]
        if existing_value is None or existing_value == "":
            existing_metadata[key] = new_value
            continue

        merged = _merge_list_and_str(existing_value, new_value)
        if merged != existing_value:
            existing_metadata[key] = merged

    return existing_metadata
