pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG QA with PDF documents"
readme = "README.md"
authors = [
    { name = "Srinivas Laxminarayan", email = "srini.lax86@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-qdrant>=0.2.1",
    "langgraph>=0.6.4",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
    "tabulate>=0.9.0",
]
